test: run the full 14-case parser suite in workerd as well
Extract the 14 parser regression cases into src/parser/server-parser-test-cases.ts
so they can be shared between test files, and invoke them from:

- src/parser/server-html-parser.test.ts   (Node + linkedom, factory direct)
- src/server.test.ts                      (Node + linkedom, via /server entry)
- src/server.worker.test.ts               (workerd + linkedom/worker)

Previously the workerd pool only ran 2 smoke cases. The full suite now
runs inside real workerd, raising confidence that linkedom/worker parses
complex inputs (tables, iframes, nested lists, media, etc.) the same
way linkedom does in Node.
